Refactor Tutorial store models
