+++ packages/lib-classifier/src/store/Medium/index.js
@@ -1,0 +1,1 @@
+export { default } from './Medium'
