+++ packages/lib-classifier/src/store/TutorialStore/Tutorial/Tutorial.js
@@ -1,5 +1,5 @@
 import { types } from 'mobx-state-tree'
-import Resource from './Resource'
+import Resource from '@store/Resource'
 
 const Tutorial = types
   .model('Tutorial', {
