+++ packages/lib-classifier/src/store/TutorialStore/Tutorial/index.js
@@ -1,0 +1,1 @@
+export { default } from './Tutorial'
