+++ packages/lib-classifier/src/store/TutorialStore/index.js
@@ -1,0 +1,1 @@
+export { default } from './TutorialStore'
